+++ src/hictk/cli/cli_validate.cpp
@@ -68,8 +68,8 @@ void Cli::transform_args_validate_subcommand() {
     c.verbosity = static_cast<std::uint8_t>(spdlog::level::err);
   } else {
     // in spdlog, high numbers correspond to low log levels
-    assert(c.verbosity > 0 && c.verbosity < 5);
-    c.verbosity = static_cast<std::uint8_t>(spdlog::level::critical) - c.verbosity;
+    assert(c.verbosity > 0 && c.verbosity < 5);  // NOLINTNEXTLINE(*-narrowing-conversions)
+    c.verbosity = static_cast<std::int16_t>(spdlog::level::critical) - c.verbosity;
   }
 }
